requirements.txt
# Bot Telegram
python-telegram-bot[webhooks]==21.10

# API
fastapi==0.115.12
uvicorn[standard]==0.34.2

# LLM
anthropic==0.52.0

# Scheduler
apscheduler==3.11.0

# Google Calendar
google-api-python-client==2.166.0
google-auth==2.40.1

# Database
supabase==2.13.0

# Rate limiting
slowapi==0.1.9

# Config
python-dotenv==1.1.0
pydantic-settings==2.9.1

# Logging
structlog==25.3.0

# HTTP client (bot -> API)
httpx==0.28.1

# Timezone
pytz==2025.2
